[physical-storage] Move Read/Write to its own file
